+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Fixed
 - Corretto errore in fase di elaborazione di un messaggio: Part NULL (Task #16775).
 - Corretto query su ricerca documento per aggancio notifiche. Ci sono casi nei quali vengono fatte ricerche su nrecord anziche' num_prot (Task #16797).
+- Corretto controllo su porta socket occupata. Su versione precedente possibilità di avviare istanze multiple di MSA.
+- Aggiunta del riferimento al thread corrente allo username dell'utente per xw per eliminazione di eventuali problemi di "Protezione file non riuscita" dovuti a gestione multithread.
 
 ## [3.0.2] - 2018-12-07